file and image upload improvements

+++ models/Trailers.php
@@ -10,7 +10,7 @@ class Trailers extends ActiveRecord{
 
     public static function deleteTrailer($id){
         $trailer = Trailers::findOne($id);
-        if($trailer->picture && file_exists(Yii::$app->request->baseUrl).'/web/images/trailers/'.$trailer->picture){
+        if($trailer->picture && $trailer->picture !== 'default.jpg' && file_exists(Yii::$app->request->baseUrl).'/web/images/trailers/'.$trailer->picture){
             unlink($_SERVER['DOCUMENT_ROOT'].Yii::$app->request->baseUrl.'/web/images/trailers/'.$trailer->picture);
         }
         return $trailer->delete();
